participants/team/source/backend/requirements.txt
Flask>=3.1,<4
Flask-SQLAlchemy>=3.1,<4
Flask-Migrate>=4.1,<5
Flask-Login>=0.6,<1
Flask-Cors>=5,<7
pydantic[email]>=2.10,<3
python-dotenv>=1.0,<2
gunicorn>=23,<26
psycopg[binary]>=3.2,<4
openpyxl>=3.1,<4
pytest>=8,<9
